src/requirements.txt
boto==2.27.0
backports.lzma==0.0.2
Flask==0.9
PyYAML==3.10
simplejson==3.1.3
requests==1.2.0
gunicorn==18.0
gevent==0.13.8
blinker==1.3
python-glanceclient==0.10.0
python-keystoneclient==0.3.1
python-swiftclient==1.8.0
redis==2.8.0
sqlalchemy==0.9.2
bugsnag==1.3.1
gcs-oauth2-boto-plugin>=1.3
bugsnag==1.3.1
M2Crypto==0.22.3
